// repositorio/SIMTR/4-implementacao/simtr-web/src/app/model/local-storage-visualizador.ts
export class LocalStorageVisualizador {
    dossieProduto?: any;
    idDocumento?: number;
    listDocumentosImagensComparar?: any[];
    visualizadorDois?: boolean;
    idVisualizar?: number;
    listaDocumentacao?: any[];
    alutraMaxHeigthImg?: number;
    idCombo?: number;
    layoutVerticalA?: boolean;
    layoutVerticalB?: boolean;
    hasMapListaCombo?;
    exibindoChecklistFase?: boolean;
    opemModal?: boolean;
    listaDocumento?: any[];
    listDocumentosImagens?: any[];
    documentoSelecionado?: any;
    chekListFase?: boolean;
}